% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - Axomatic(Conflicted).r, R/script -
%   Axomatic.r
\name{axis_y}
\alias{axis_y}
\title{axis_y}
\usage{
axis_y(
  from,
  to,
  ticks = 1,
  labels = 1,
  pad = 0,
  upperPad = NA,
  lowerPad = NA,
  expand = 0,
  upperExp = NA,
  lowerExp = NA,
  trans = "identity",
  trans_labels = "none"
)

axis_y(
  from,
  to,
  ticks = 1,
  labels = 1,
  pad = 0,
  upperPad = NA,
  lowerPad = NA,
  expand = 0,
  upperExp = NA,
  lowerExp = NA,
  trans = "identity",
  trans_labels = "none"
)
}
\arguments{
\item{from}{Start of the axis.}

\item{to}{End of the axis.}

\item{ticks}{How much space between each tickmark.}

\item{labels}{The numbers or text beneath the tickmarks. 0 = no labels, 1 = label on every tickmark, 5 = label every 5'th tickmark, and so on. Setting this paramter as a vector lets you make self-defined labels, e.g. you can write c("one","two","three","four") or whatever.}

\item{pad}{How much space before and after the first and last mark on the axis.}

\item{upperPad}{specification of pad, but only for space after the last mark.}

\item{lowerPad}{specifiction of pad, but only for spcae before the first mark.}

\item{trans_labels}{"none": Does nothing (default). Lognatural: Transforms numerical labels back to non-logarithmic form if the axis data is log transformed}
}
\description{
Simplified function for defining the y axis with settings for tickmarks and labels.
The function is used as a ggplot object.

Simplified function for defining the y axis with settings for tickmarks and labels.
The function is used as a ggplot object.
}
\examples{
axis_y(from=0, to=20, ticks=1, labels=5, pad=2) #makes an axis going from 0 to 20, with 20 tickmarks, and labels every 5th tickmark, and a space of 2 before and after te marks.
axis_y(from=0, to=20, ticks=1, labels=5, pad=2) #makes an axis going from 0 to 20, with 20 tickmarks, and labels every 5th tickmark, and a space of 2 before and after te marks.
}
