% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - Axomatic.r
\name{axomatic}
\alias{axomatic}
\title{Axomatix}
\usage{
axomatic(x, y, ratio = 1, scaleSize = T)
}
\arguments{
\item{x}{an x-axis specified with the function axis_x (or aximaticultimatic())}

\item{y}{an y-axis specified with the function axis_y (or aximaticultimatic())}

\item{ratio}{the ratio between the x and y axis. Defaults to 1 (a square)}

\item{scaleSize}{if the ratio is for physical graph size (T, default) or the va}
}
\description{
Used to make defining x axis, y axis, and plot shape (square, rectangular, etc)
It uses functions axis_x() and axis_y() to make two axes, and then lets the user specify
the shape of the plot.
The function is used and added as a ggplot object.
}
\examples{
axomatic(x_axis(0,20),y_axis(1000, 10000))
}
