% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - Axomatic.r
\name{base_axis}
\alias{base_axis}
\title{base_axis}
\usage{
base_axis(
  xy,
  from,
  to,
  ticks = 1,
  labels = 1,
  pad = 0,
  upperPad = NA,
  lowerPad = NA,
  trans = "identity",
  trans_labels = "none"
)
}
\arguments{
\item{xy}{Whether this is a x ("x") axis or y axis ("y").}

\item{from}{Start of the axis.}

\item{to}{End of the axis.}

\item{ticks}{How much space between each tickmark.}

\item{labels}{The numbers or text beneath the tickmarks. 0 = no labels, 1 = label on every tickmark, 5 = label every 5'th tickmark, and so on. Setting this paramter as a vector lets you make self-defined labels, e.g. you can write c("one","two","three","four") or whatever.}

\item{pad}{How much space before and after the first and last mark on the axis.}

\item{upperPad}{specification of pad, but only for space after the last mark.}

\item{lowerPad}{specifiction of pad, but only for spcae before the first mark.}

\item{trans_labels}{"none": Does nothing (default). Lognatural: Transforms numerical labels back to non-logarithmic form if the axis data is log transformed}
}
\description{
Internal function that's gives axis_y() and axis_x() their base functionality.
The function is used as a ggplot object.
}
\examples{
axis_x(from=0, to=20, ticks=1, labels=5, pad=2) #makes an axis going from 0 to 20, with 20 tickmarks, and labels every 5th tickmark, and a space of 2 before and after te marks
}
