% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - Axomatic(Conflicted).r, R/script -
%   Axomatic.r
\name{numextract}
\alias{numextract}
\title{Converts anything to a numeric}
\usage{
numextract(string)

numextract(string)
}
\description{
Converts anything to a numeric

Converts anything to a numeric
}
