% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script - Axomatic.r
\name{axomatic}
\alias{axomatic}
\title{Axomatic}
\usage{
axomatic(x, y, ratio = 1, scaleSize = T)
}
\arguments{
\item{x}{an x-axis specified with the function axis_x}

\item{y}{an y-axis specified with the function axis_y}

\item{ratio}{the ratio between the x and y axis. Defaults to 1 (a square)}

\item{scaleSize}{if the ratio is for physical graph size (T, default) or the values displayed. If physical (T), setting ration to 1 will always result in a square graph. If not physical (F), setting ration to 1 will give equal scale on both axes.}
}
\description{
Used for defining x axis, y axis, and plot shape (square, rectangular, etc)
It uses functions axis_x() and axis_y() to make two axes, and then lets the user specify
the shape of the plot.
The function is used and added as a ggplot object.
See help page for axis_x() and axis_y().
}
\examples{
axomatic(x_axis(from=0,to=20,ticks=5,labels=1),y_axis(1000, 10000, 100, 2))
}
